Move parse_dice_code to Dragon superclass
